requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7
